QCamera2: Change focus num steps type
Change the type to int_32 as focus steps can go upto 1024

+++ camera/QCamera2/stack/common/cam_types.h
@@ -707,7 +707,7 @@ typedef struct {
 typedef struct {
     uint8_t stepsize;
     uint8_t direction;
-    uint8_t num_steps;
+    int32_t num_steps;
     uint8_t ttype;
 } tune_actuator_t;
